Fixing the DOCA release build by cleaning up the includes
