+++ morpheus/_lib/doca/include/morpheus/doca/doca_mem.hpp
@@ -17,8 +17,8 @@
 
 #pragma once
 
-#include "doca_context.hpp"
-#include "error.hpp"
+#include "morpheus/doca/doca_context.hpp"
+#include "morpheus/doca/error.hpp"
 
 #include <doca_gpunetio.h>
