+++ morpheus/_lib/doca/include/morpheus/doca/doca_rx_queue.hpp
@@ -17,41 +17,35 @@
 
 #pragma once
 
-#include "common.hpp"
-#include "error.hpp"
-#include "rte_context.hpp"
+#include "morpheus/doca/doca_context.hpp"
+#include "morpheus/doca/doca_mem.hpp"
 
 #include <doca_eth_rxq.h>
-#include <doca_flow.h>
 #include <doca_gpunetio.h>
-#include <doca_log.h>
-
-#define GPU_PAGE_SIZE (1UL << 16)
 
 namespace morpheus::doca {
 
 /**
- * @brief Manages the lifetime of DOCA as it relates to GPUNetIO
+ * @brief Creates and manages the lifetime of a GPUNetIO Receive Queue.
+ *
+ * A Receive Queue is used to buffer packets received from a GPUNetIO Pipe.
  */
-struct DocaContext
+struct DocaRxQueue
 {
   private:
-    doca_gpu* m_gpu;
-    doca_dev* m_dev;
-    doca_flow_port* m_flow_port;
-    uint16_t m_nic_port;
-    uint32_t m_max_queue_count;
-    std::unique_ptr<RTEContext> m_rte_context;
-    doca_log_backend* sdk_log;
+    std::shared_ptr<DocaContext> m_context;
+    doca_gpu_eth_rxq* m_rxq_info_gpu;
+    doca_eth_rxq* m_rxq_info_cpu;
+    doca_mmap* m_packet_mmap;
+    doca_ctx* m_doca_ctx;
+    std::unique_ptr<DocaMem<void>> m_packet_memory;
 
   public:
-    DocaContext(std::string nic_addr, std::string gpu_addr);
-    ~DocaContext();
+    DocaRxQueue(std::shared_ptr<DocaContext> context);
+    ~DocaRxQueue();
 
-    doca_gpu* gpu();
-    doca_dev* dev();
-    uint16_t nic_port();
-    doca_flow_port* flow_port();
+    doca_gpu_eth_rxq* rxq_info_gpu();
+    doca_eth_rxq* rxq_info_cpu();
 };
 
 }  // namespace morpheus::doca
